Create Resume SNR
The SANTRAST "SNR" token has a basic yield, is the main asset, accumulates liquidity due to cross exchange rate recalculations of DeFi protocols. Interacts with the SNRW contract and the DEX contract in which it is placed. It is designed to hedge positions, increase capital due to the AMM multiplier and its isolation. The result of such a strategy is visible during the period of mass sales, its liquidity does not change. During periods of growth of the crypto market, SNR increases its liquidity, even if trading is not conducted on it.